ai_wealth_manager/requirements.txt
streamlit
crewai
crewai-tools
langchain-google-genai
numpy
scipy
plotly
python-dotenv
pandas
